refactor: config keys.
