+++ packages/next/src/sitemap/services.ts
@@ -1,7 +1,7 @@
 import { XMLParser, type X2jOptions } from 'fast-xml-parser';
 import { Logger } from '@snapwp/core';
 import { getConfig } from '@snapwp/core/config';
-import type { SitemapDataFromXML } from '@/sitemap/utils';
+import type { SitemapDataFromXML } from '@snapwp/types';
 
 const XMLParserConfig: Partial< X2jOptions > = {
 	ignoreAttributes: false,
@@ -21,13 +21,13 @@ const XMLParserConfig: Partial< X2jOptions > = {
 export const fetchIndexSitemap = async (
 	customXMLParserConfig?: X2jOptions
 ): Promise< SitemapDataFromXML[] > => {
-	const { wpHomeUrl, sitemapConfig } = getConfig();
+	const { wpHomeUrl, sitemap: sitemapConfig } = getConfig();
 
-	if ( ! sitemapConfig?.indexSitemapUri ) {
+	if ( ! sitemapConfig?.indexUri ) {
 		return [];
 	}
 
-	const wpSitemapIndexPath = `${ wpHomeUrl }/${ sitemapConfig.indexSitemapUri }`;
+	const wpSitemapIndexPath = `${ wpHomeUrl }/${ sitemapConfig.indexUri }`;
 
 	const response = await fetch( wpSitemapIndexPath );
