+++ packages/types/src/index.ts
@@ -1,4 +1,4 @@
 export * from './blocks';
-export * from './utils';
-export * from './sitemap';
 export * from './query-engine';
+export * from './sitemap';
+export * from './utils';
