+++ packages/types/src/sitemap/index.ts
@@ -1,7 +1,7 @@
 import type { MetadataRoute } from 'next';
 
 export type SitemapConfig = {
-	indexSitemapUri?: string;
+	indexUri?: string;
 	ignorePatterns?: string[];
 	customPaths?: SitemapCustomPath;
 };
@@ -13,4 +13,19 @@ export type SitemapData = Omit<
 
 export type SitemapCustomPath = {
 	[ subType: string ]: SitemapData[];
+};
+
+export type SitemapDataFromXML = {
+	loc?: string;
+	lastmod?: string | Date | undefined;
+	changeFrequency?:
+		| 'always'
+		| 'hourly'
+		| 'daily'
+		| 'weekly'
+		| 'monthly'
+		| 'yearly'
+		| 'never'
+		| undefined;
+	priority?: number | undefined;
 };
